backend/requirements.txt
asgiref==3.11.1
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
packaging==26.0
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dotenv==1.2.2
ruff==0.15.8
sqlparse==0.5.5
tzdata==2025.3
bandit==1.9.4
safety==3.7.0
pytest-django==4.9.0
